REST Controllers: Make parent post data embeddable
For both the `get_items` and `get_item` endpoints, this adds the option
of embedding data about the parent post(s) in the response by adding
a `_embed=true` parameter to the request.

The use case is getting access to the parent post's post type, but there
could be other applications as well.