Add a new screen to compare a revision to its parent post (#43)
The built-in revision.php screen in the WP admin is too opinionated and not flexible enough to show things the way we want. This custom screen offers a simplified version, without the buttons and checkboxes, the restore feature, and the slider bar to switch between different revisions. The intention here is just to show how one revision compares to the current post.

Also consolidates how assets are enqueued in the admin.

Refs #5