Handle cron events for scheduled updates (#45)
This works through the same system that schedules events to publish posts that get saved with the `future` post status. First we have to remove the event that WP creates, and then add our own that will run `update_post_from_revision` instead of `wp_update_post`. This also handles some, but not all of the scenarios where an update should no longer be applied at its specified future date, so the event gets deleted.