Check for and delete an autosave after successful update creation
It appears that deleting the autosave during the `create_item` request
doesn't prompt the editor to try to "resave" again later, so there isn't
any need to try and prevent another autosave from happening if the user
hangs out with the modal for a while after creating an update.

Implementing this in the REST controller instead of in
`put_post_revision` because it seems like the appropriate "distance"
from the side effects that manifest from having an autosave that matches
the update content. The `put_post_revision` PHP function is too low
level, and should only be concerned with storing a revision, not other
moving pieces around it.

Fixes #17