Handle UX for updates of posts that are no longer published (#60)
This simply changes how updates are presented (and how their parent posts are presented) in the WP Admin when the parent post is not published. If we modified these updates when the parent status changed (e.g. assigned a different post status), we would then have to modify the updates again when the parent is republished, and both of those processes would be fragile and potentially miss edge cases. Fortunately, updates are pretty harmless the way they are. If a scheduled update's cron event triggers, and the parent post is not published, the cron silently fails, and the update continues on with a future status, but no further cron job scheduled.

Fixes #37