Add php unit test suite (#80)
Sets up PHP unit testing, and adds tests for the revisions controller. Tests for other parts of the code can be added in separate PRs.

In creating the revisions controller tests, a couple of minor changes were also made to the controller code itself.

Note that this required changes to the wp-env config. Setting the php version to 5.6 seemed like a good idea since we're trying to develop to the same minimum requirements as Core, but there are bugs in wp-env that prevent things from working right when a custom version is set.

See:
https://github.com/WordPress/gutenberg/issues/29323
https://github.com/WordPress/gutenberg/issues/27783

Fixes #71