Try: Split dropdown button with options. (#94)
This changes the flow and UX for creating a new update of a post. This new flow is more "self-documenting" because it's discoverable during the default flow of saving changes to a post via a button dropdown, and then the sidebar that pops out provides the additional context to understand the choices available.

This also moves the "Publish immediately" function in the revision editor into a similar button dropdown.

This is a somewhat rough v1. Enhancements to this can happen in separate PRs.

Fixes #87